confirmation dialog moved to components in sections
