+++ Web Ui/src/sections/Assignments/components/ConfirmationDialog.tsx
